src/sub-agent/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
redis>=5.0.0
httpx>=0.24.0
python-dotenv>=1.0.0
structlog>=23.1.0
opentelemetry-api>=1.18.0
opentelemetry-sdk>=1.18.0
opentelemetry-instrumentation-fastapi>=0.39b0
opentelemetry-exporter-otlp>=1.18.0
pyyaml>=6.0
langgraph>=0.0.10
langchain>=0.0.300
langchain-openai>=0.1.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
fakeredis[lua]>=2.14.1
